requirements.txt
pandas
numpy
matplotlib
scikit-learn
seaborn
langchain
faiss-cpu
langchain
langchain-community
langchain-core
langchain-groq
langchain-openai
langchain-huggingface
langchain-google-genai==2.1.8
chromadb
pypdf
openai
groq
streamlit
faiss-cpu
PyMuPDF
networkx
python-dotenv
ipykernel
structlog
